+++ cmd/lfgw/main.go
@@ -13,6 +13,7 @@ import (
 	oidc "github.com/coreos/go-oidc/v3/oidc"
 	"github.com/rs/zerolog"
 	zlog "github.com/rs/zerolog/log"
+	"github.com/weisdd/lfgw/internal/acl"
 	"go.uber.org/automaxprocs/maxprocs"
 )
 
@@ -21,7 +22,7 @@ import (
 type application struct {
 	errorLog                *log.Logger
 	logger                  *zerolog.Logger
-	ACLMap                  ACLMap
+	ACLMap                  acl.ACLMap
 	proxy                   *httputil.ReverseProxy
 	verifier                *oidc.IDTokenVerifier
 	Debug                   bool          `env:"DEBUG" envDefault:"false"`
@@ -98,7 +99,7 @@ func main() {
 	}
 
 	if app.ACLPath != "" {
-		app.ACLMap, err = app.loadACL(app.ACLPath)
+		app.ACLMap, err = acl.NewACLMapFromFile(app.ACLPath)
 		if err != nil {
 			app.logger.Fatal().Caller().
 				Err(err).Msgf("Failed to load ACL")
